requirements.txt
flask==2.0.1
psycopg2==2.9.1
python-decouple